Added confirmation for deleting cards and added some other transform for bigger screens than 400px

+++ src/pages/Home.jsx
@@ -29,7 +29,9 @@ const Home = ({ cards, setCards, activeCard, setActiveCard }) => {
             <button
               className="active-card__button-delete"
               onClick={() => {
-                deleteCard(activeCard);
+                if (confirm("Do you want to delete this card?")) {
+                  deleteCard(activeCard);
+                }
               }}
             >
               Delete Card
